refactor: relocate _build_adapter_policy to map_runner_policy_common (#3403)
Prerequisite for the #3384 adapter-family decomposition. _build_adapter_policy
is the shared helper that ~17 of _build_policy's adapter branches funnel through,
but it lived inside map_runner.py — so migrating those branches into
map_runner_policies/ builder modules would force an import cycle back into
map_runner.

- Move the helper verbatim into robot_sf/benchmark/map_runner_policy_common.py as
  build_adapter_policy(), importing its dependencies from their canonical modules
  (planner_command_contract, algorithm_metadata, utils, map_runner_policy_metadata,
  planner.kinematics_model) — none of which import map_runner, so the new module is
  acyclic.
- Re-export it into map_runner as `_build_adapter_policy` so all existing call sites
  are unchanged.
- Behavior-preserving: the map_runner regression suite (test_map_runner_utils +
  ppo/sac/lidar/policy_stack/crowdnav/social_force) passes unchanged. The three
  resolve_benchmark_kinematics_model monkeypatches in tests target _ppo_action_to_unicycle
  and the goal path (not _build_adapter_policy), so they are unaffected.
